pyproject.toml
[tool.poetry]
name = "cupas"
version = "0.1.0"
description = ""
authors = ["pftrobot <pftrobot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "2.32.3"
annotated-types = "0.7.0"
anyio = "4.4.0"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.7.4"
cffi = "1.17.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
distlib = "0.3.8"
distro = "1.9.0"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jiter = "0.5.0"
keyring = "24.3.1"
more-itertools = "10.4.0"
msgpack = "1.0.8"
openai = "1.42.0"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pyproject-hooks = "1.1.0"
python-dotenv = "1.0.1"
rapidfuzz = "3.9.6"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
tomlkit = "0.13.2"
tqdm = "4.66.5"
trove-classifiers = "2024.7.2"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
xattr = "1.1.0"
beautifulsoup4 = "^4.12.3"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
selenium = "^4.24.0"
webdriver-manager = "^4.0.2"
cv = "^1.0.0"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
